When the queue gets large, push to disk instead.
We were expereancing starling processes inflating to 1gb or more when a substantial number of messages were posted. This caused swap which caused lag, which cause pollers to time out, which caused queues to grow. Eventualy this caused server death.

With this change, we only keep the first 10,000 messages in memory (enough for normal operations) and in a faling state we keep the rest on disk. Batches of 10,000 are loaded from disk once the main queue is empty.